Fix up/down arrow key behavior on steevie
I have no idea why this is necessary but long ago I had an issue and
this seemed to fix it. So then it sat uncommitted for a long while
until I got tired of feeling bad that git status constantly said I had
a dirty working directory, and decided to do the right thing and commit
it.

Better late than never.