Clean up the ZNC code
And at the same time, add the last part of the lambda fix mentioned in
df7829e23257500fc8c48a301c09334f9c4f39ac - (if (not)) -> (unless).

We also use (auth-source-creation-defaults) since that was in the
sample code in the (auth-source-search) docs and it seems like a good idea.